changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11824 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/HV/UnivVoltHVCrate/ORUnivVoltHVCrateModel.m
@@ -1051,7 +1051,7 @@ NSString* HVkErrorMsg = @"ErrorMsg";
 }
 
 
-- (void) netsocket: (NetSocket*) anInNetSocket dataAvailable: (unsigned) anInAmount
+- (void) netsocket: (NetSocket*) anInNetSocket dataAvailable: (NSUInteger) anInAmount
 {
     @try {
 		if (anInNetSocket == mSocket) {
